Carry out a Pearson function to particular package
